core: Use rustfft for the MDCT.
RustFFT provides safe SIMD accelerated transforms. A semver breaking
change is required to switch the backend implementation of the fft
module, however, the imdct module can use it now.

In v0.6.0, the fft module will use rustfft as a backend, and the
complex module will be replaced by num_complex.